+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-03 12:00:27
-// Commit: style(state): harmonize light and dark theme palette for state
+// Updated at: 2026-07-03 13:13:40
+// Commit: test(state): implement widget integration tests for state
 
 class stateComponentState {
-  final String commitRef = '2518';
-  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
-  final DateTime timestamp = DateTime.parse('2026-07-03T12:00:27+05:00');
+  final String commitRef = '2524';
+  final String lastMessage = 'test(state): implement widget integration tests for state';
+  final DateTime timestamp = DateTime.parse('2026-07-03T13:13:40+05:00');
 }